+++ NpgsqlRest/NpgsqlRestLogger.cs
@@ -105,6 +105,9 @@ public static partial class Log
 
     [LoggerMessage(Level = LogLevel.Information, Message = "{type} {schema}.{name} has set NEW LINE by the comment annotation to {value}.")]
     public static partial void CommentSetRawNewLineSeparator(this ILogger logger, RoutineType type, string schema, string name, string value);
+
+    [LoggerMessage(Level = LogLevel.Information, Message = "{type} {schema}.{name} has set COLUMN NAMES by the comment annotation.")]
+    public static partial void CommentRawSetColumnNames(this ILogger logger, RoutineType type, string schema, string name);
 }
 
 public static class NpgsqlRestLogger