+++ plugins/NpgsqlRest.TsClient/TsClientOptions.cs
@@ -18,7 +18,8 @@ public class TsClientOptions(
     string[]? skipFunctionNames = null,
     string[]? skipPaths = null,
     string defaultJsonType = "string",
-    bool useRoutineNameInsteadOfEndpoint = false)
+    bool useRoutineNameInsteadOfEndpoint = false,
+    bool exportUrls = false)
 {
     /// <summary>
     /// File path for the generated code. Set to null to skip the code generation. Use {0} to set schema name when BySchema is true
@@ -122,4 +123,9 @@ public class TsClientOptions(
     /// Use routine name instead of endpoint name when generating a function name.
     /// </summary>
     public bool UseRoutineNameInsteadOfEndpoint { get; set; } = useRoutineNameInsteadOfEndpoint;
+
+    /// <summary>
+    /// Export URLs as constants.
+    /// </summary>
+    public bool ExportUrls { get; set; } = exportUrls;
 }